refactor(2d): extract curve base component
